always share our profileKey on outgoing messages
